Add NewTimer to tcpip.Clock
Changes the API of tcpip.Clock to also provide a method for scheduling and
rescheduling work after a specified duration. This change also implements the
NewTimer method for existing implementations of tcpip.Clock.

This is the groundwork required to mock time within tests. All references to
CancellableTimer has been replaced with the tcpip.Timer interface, allowing for
custom implementations of a timer.

This is a BREAKING CHANGE for clients that implement their own tcpip.Clock.
Additional work is required by these clients to implement tcpip.Timer and add
the NewTimer function to tcpip.Clock.

PiperOrigin-RevId: 316767762